refactor: rename proof module to nullifier_proof
