+++ crates/core/src/lib.rs
@@ -28,7 +28,7 @@ pub use world_id_issuer::Issuer;
 pub use world_id_primitives::Signer;
 
 #[cfg(feature = "authenticator")]
-pub use world_id_proof::{artifacts, proof};
+pub use world_id_proof::{artifacts, nullifier_proof};
 
 #[cfg(any(feature = "authenticator", feature = "rp"))]
 pub use world_id_primitives::request as requests;
