+++ crates/proof/src/lib.rs
@@ -20,8 +20,8 @@ pub use compress::ProofCompression;
 pub(crate) mod oprf_query;
 pub use oprf_query::{FullOprfOutput, OprfEntrypoint};
 
-pub mod proof;
-pub use proof::*;
+pub mod nullifier_proof;
+pub use nullifier_proof::*;
 
 use provekit_common::{InputMap, InputValue, NoirElement};
