+++ crates/proof/src/nullifier_proof.rs
