+++ services/oprf-dev-client/src/lib.rs
@@ -13,7 +13,7 @@ use taceo_oprf::{core::oprf::BlindingFactor, dev_client::DevClientConfig};
 use world_id_core::{
     Authenticator, AuthenticatorError, EdDSAPrivateKey, EdDSASignature, FieldElement,
     artifacts::ZkArtifactSourceExt as _,
-    proof::{CircomGroth16Material, errors},
+    nullifier_proof::{CircomGroth16Material, errors},
 };
 use world_id_primitives::{AuthenticatorPublicKeySet, TREE_DEPTH, merkle::MerkleInclusionProof};
 use world_id_proof::circuit_inputs::QueryProofCircuitInput;
